MAINT: Clean up some util functions
